+++ CHANGELOG.md
@@ -1,3 +1,38 @@
-# CloudBlue Connect Transformations Library changelog
+# CloudBlue Connect Standard Transformations Library
 
-* 27.4: initial version
+
+## 27.10
+
+* New **Lookup Data from AirTable** transformation.
+* Improve **Formula** parsing to extract input columns.
+* Allow auto and 1 decimal position for **Formula** transformation.
+* Fix **Lookup Data from a stream attached Excel** to use output column type auto detection.
+* Improve **Filter Rows by Condition** transformation to allow multiple values and match/mismatch condition.
+
+## 27.9 
+
+* New **Lookup Data from a stream attached Excel** transformation.
+
+## 27.8
+
+* Add support to specify precisions for decimal output columns for **Split Column** transformation.
+* Fix **Formula** input columns list extraction.
+
+## 27.7
+
+* Support explicit output value type casting for **Formula** and **Split Column** transformations.
+* Add input column suggest feature to **Formula** transformation configuration UI.
+
+
+## 27.6
+
+* Several minor fixes of **Formula** and **Copy Column(s)** transformations.
+
+## 27.5
+
+* New **Filter Rows by Condition** transformation.
+* Improvements of **Split Column** transformation.
+
+## 27.4
+
+First public release.